Add logic for determining when a notification should be raised
There are some hard coded parameters in here at the moment
